Clean up scorer files
