+++ cyslf/scorers/__init__.py
@@ -1,0 +1,4 @@
+from .core import DEFAULT_WEIGHTS, CompositeScorer
+
+
+__all__ = ["DEFAULT_WEIGHTS", "CompositeScorer"]
